Move snapshot widget data to separate files
