+++ RunRideStudioWidget/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
