+++ RunRideStudioWidget/Resource/Assets.xcassets/Contents.json
