+++ RunRideStudioWidget/Resource/Assets.xcassets/WidgetBackground.colorset/Contents.json
