+++ RunRideStudioWidget/Resource/Info.plist
