+++ RunRideStudioWidget/Resource/RunRideStudioWidgetExtension.entitlements
