+++ RunRideStudioWidget/Source/RunRideStudioWidgetBundle.swift
@@ -12,7 +12,7 @@ import WidgetKit
 struct RunRideStudioWidgetBundle: WidgetBundle {
     @WidgetBundleBuilder
     var body: some Widget {
-        Snapshot()
+        SnapshotWidget()
         RunRideStudioWidget()
     }
 }
